remove main/friend container
