+++ packages/client/src/containers/index.tsx
@@ -1,3 +1,1 @@
-export { default as FriendContainer } from './FriendContainer'
-export { default as MainContainer } from './MainContainer'
 export { default as ChatRoomStartContainer } from './ChatRoomStartContainer'
