+++ packages/client/src/pages/main/index.tsx
@@ -1,27 +1,34 @@
-import React, { FC } from 'react'
+import React, {
+  FC, useEffect,
+} from 'react'
 import * as S from 'pages/main/styles'
 import NavigationBar from 'components/NavigationBar'
-import { RootState } from 'modules'
-import { Loader } from 'components'
-import FriendContainer from 'containers/FriendContainer'
 import Room from 'system/Room'
-import { useSelector } from 'react-redux'
+import { useDispatch } from 'react-redux'
 import { Route } from 'react-router-dom'
 import { url } from 'common/constants'
+import { getRoomRequest } from 'modules/room'
+import { afterLogin } from 'socket'
+import Friend from 'pages/main/Friend'
+import { getUserInfo } from 'common/request'
 
 const Main: FC = () => {
-  const { isLoggedIn } = useSelector((state: RootState) => state.login)
+  const dispatch = useDispatch()
+  useEffect(() => {
+    dispatch(getRoomRequest())
+    getUserInfo().then((response) => {
+      const { uuid } = response.data.data
+      afterLogin({ uuid })
+    })
+  })
+  window.history.pushState(null, '', window.location.href)
 
   return (
     <S.Container>
-      {isLoggedIn ? (
         <S.MainWrapper>
-          <Route path={url.main.friendList} component={FriendContainer}/>
+          <Route path={url.main.friendList} component={Friend}/>
           <Route path={url.main.chatList}component={Room}/>
         </S.MainWrapper>
-      ) : (
-        <Loader/>
-      )}
       <S.NavigationBarWrapper>
         <NavigationBar/>
       </S.NavigationBarWrapper>
